pipeline/pyproject.toml
[project]
name = "thirukkural-pipeline"
version = "0.1.0"
description = "Offline generation pipeline: kural -> story + 4-panel folk-art comic."
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "google-genai>=0.8.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = false

[tool.pyright]
include = ["."]
extraPaths = ["."]
reportMissingImports = "warning"
